Separate capabilities and processes in different files
