+++ src/ast/mod.rs
@@ -1,1 +1,2 @@
-pub mod term;
+pub mod capabilities;
+pub mod processes;
